requirements.txt
asgiref==3.2.7
blessings==1.7
bpython==0.19
certifi==2020.4.5.1
chardet==3.0.4
curtsies==0.3.1
Django==3.0.5
djangorestframework==3.11.0
django-cors-headers==3.2.1
greenlet==0.4.15
honcho==1.0.1
idna==2.9
Markdown==3.2.1
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
wcwidth==0.1.9
